feature-matching-api/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Computer Vision and Image Processing
opencv-python-headless==4.8.1.78
numpy==1.24.4  # Exact version to avoid NumPy 2.x compatibility issues

# Optional GeoTIFF support
rasterio==1.3.9
pyproj==3.6.1

# Geometry processing
shapely==2.0.2

# Map cutting dependencies
requests==2.31.0
tqdm==4.66.1

# Machine Learning and Deep Learning
torch==2.0.1
torchvision==0.15.2

# GIM Roma dependencies
huggingface_hub==0.17.3
einops==0.7.0
timm==0.9.7
kornia==0.7.0

# Additional utilities
python-json-logger==2.0.7
Pillow==10.0.1
psutil==5.9.6

# Development and testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1 